[level 0] Title: 두 수의 나눗셈, Time: 0.00 ms, Memory: 10.1 MB -BaekjoonHub

+++ 프로그래머스/lv0/120806. 두 수의 나눗셈/두 수의 나눗셈.py
@@ -1,4 +1,5 @@
 def solution(num1, num2):
-    answer = ((num1 / num2) * 1000) // 1
-    
+    # answer = ((num1 / num2) * 1000) // 1
+    # 정수 반환
+    answer = int(num1 / num2 * 1000)
     return answer
